Feat: 프로젝트 상세 페이지 추가

+++ src/main/java/com/project/toyple/project/ProjectDao.java
@@ -3,7 +3,6 @@ package com.project.toyple.project;
 import org.springframework.data.jpa.repository.JpaRepository;
 
 //인터페이스 선언만으로 jpa관련된 많은 메서드 활용 가능
-public interface ProjectDao extends JpaRepository<ProjectDto, Long> {
-
+public interface ProjectDao extends JpaRepository<ProjectDto, Integer> {
 
 }
